More more functions to utils.py
